+++ sc/social/like/plugins/linkedin/browser.py
@@ -1,57 +1,53 @@
 # -*- coding:utf-8 -*-
+"""Helper view to generate LinkedIn widget.
+
+More information:
+* https://developer.linkedin.com/plugins/share
+* https://developer.linkedin.com/docs/share-on-linkedin
+"""
+from Acquisition import aq_inner
 from plone import api
-from plone.api.exc import InvalidParameterError
 from Products.Five import BrowserView
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from sc.social.like.interfaces import ISocialLikeSettings
 from sc.social.like.utils import get_language
 from urllib import urlencode
-from zope.component import getMultiAdapter
 
 
 class PluginView(BrowserView):
+    """Helper view to generate LinkedIn widget."""
 
-    linkedin_enabled = True
-    language = 'en'
-
-    metadata = ViewPageTemplateFile('templates/metadata.pt')
+    metadata = None
     plugin = ViewPageTemplateFile('templates/plugin.pt')
     link = ViewPageTemplateFile('templates/link.pt')
 
     def __init__(self, context, request):
-        self.context = context
+        self.context = aq_inner(context)
         self.request = request
-        # FIXME: the following could rise unexpected exceptions
-        #        move it to a new setup() method
-        #        see: http://docs.plone.org/develop/plone/views/browserviews.html#creating-a-view
-        self.portal_state = getMultiAdapter((self.context, self.request),
-                                            name=u'plone_portal_state')
-        self.portal = self.portal_state.portal()
-        self.site_url = self.portal_state.portal_url()
-        self.portal_title = self.portal_state.portal_title()
-        self.url = context.absolute_url()
-        self.language = get_language(context)
 
-    @property
-    def typebutton(self):
+    def portal_url(self):
+        portal = api.portal.get()
+        return portal.absolute_url()
+
+    def language(self):
+        # XXX: future use
+        return get_language(self.context)
+
+    def counter(self):
+        """Return plugin count mode."""
         record = ISocialLikeSettings.__identifier__ + '.typebutton'
-        try:
-            typebutton = api.portal.get_registry_record(record)
-        except InvalidParameterError:
-            typebutton = ''
+        typebutton = api.portal.get_registry_record(record, default='')
 
-        if typebutton == 'horizontal':
-            typebutton = 'right'
-        else:
-            typebutton = 'top'
-        return typebutton
+        if typebutton == 'vertical':
+            return 'top'
+        return 'right'
 
     def share_link(self):
-        params = dict(
-            mini='true',
-            url=self.context.absolute_url(),
-            title=self.context.Title(),
-            summary=self.context.Description(),
-        )
+        params = {
+            'mini': 'true',
+            'url': self.context.absolute_url(),
+            'title': self.context.Title(),
+            'summary': self.context.Description(),
+        }
         url = 'https://www.linkedin.com/shareArticle?' + urlencode(params)
         return url