+++ CHANGES.rst
@@ -6,6 +6,9 @@ There's a frood who really knows where his towel is.
 2.10.2 (unreleased)
 ^^^^^^^^^^^^^^^^^^^
 
+- Update LinkedIn's Share Plugin code; code clean up and refactor.
+  [hvelarde]
+
 - Update Tweet Button code; code clean up and refactor.
   [hvelarde]